backend/requirements.txt
# dependencies
beautifulsoup4
colorama
md2pdf
python-dotenv
pyyaml
uvicorn
pydantic[email]>=2.0.0
email-validator
fastapi
python-multipart
markdown
langchain>=0.2,<0.3
langchain_community>=0.2,<0.3
langchain-openai>=0.1,<0.2
langgraph
tiktoken
gpt-researcher
arxiv
PyMuPDF
requests
jinja2
aiofiles
mistune
python-docx
htmldocx
lxml_html_clean
websockets
unstructured
json_repair
json5
loguru
browser_cookie3
firebase-admin>=6.6.0
stripe

# File format conversion
weasyprint>=60.1
python-docx>=0.8.11
Markdown>=3.4.3
markdown-extra>=0.5.0

# uncomment for testing
# pytest
# pytest-asyncio
